requirements.txt
streamlit>=1.36.0
pydantic>=2.7.0
pyyaml>=6.0.1
